refactor: split Auth middleware in JWT and API token files
