+++ crate/server/src/config/command_line/http_config.rs
@@ -31,8 +31,9 @@ pub struct HttpConfig {
     #[clap(long, env = "KMS_AUTHORITY_CERT_FILE")]
     pub authority_cert_file: Option<PathBuf>,
 
-    #[clap(long, env = "KMS_AUTHENTICATION_TOKEN")]
-    pub authentication_token: Option<String>,
+    /// The API token to use for authentication
+    #[clap(long, env = "KMS_API_TOKEN")]
+    pub api_token: Option<String>,
 }
 
 impl Display for HttpConfig {
@@ -68,7 +69,7 @@ impl Default for HttpConfig {
             https_p12_file: None,
             https_p12_password: None,
             authority_cert_file: None,
-            authentication_token: None,
+            api_token: None,
         }
     }
 }
