+++ crate/server/src/middlewares/mod.rs
@@ -1,5 +1,8 @@
+mod main;
+pub(crate) use main::AuthTransformer;
+
 mod jwt_auth;
-pub(crate) use jwt_auth::{JwtAuth, JwtAuthClaim};
+pub(crate) use jwt_auth::{manage_jwt, manage_jwt_request, JwtAuthClaim};
 
 mod ssl_auth;
 pub(crate) use ssl_auth::{extract_peer_certificate, PeerCommonName, SslAuth};
@@ -9,3 +12,6 @@ pub(crate) use jwt::{JwtConfig, JwtTokenHeaders, UserClaim};
 
 mod jwks;
 pub(crate) use jwks::JwksManager;
+
+mod token_auth;
+pub(crate) use token_auth::{manage_token, manage_token_request};
